requirements.txt
requests
simplejson
setuptools
pyyaml
pathlib